first bullet and timed cannons
created a list of weapons

added vars to the weapon class

used an interval to call update weapons, this takes weapons and sees if they should be destroyed or created this is so not every weapon is active at the same time.

added the first bullet type

TODO randomly select a position out of the cannon zones for the weapons to spawn at, weapons cannot spawn at zones that are in use.

TODO add hitboxes for the bullets to be destoryed when they hit the map bounds or the player

TODO add a way for the player to go down or up levels